requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
asgiref=3.3.4=pypi_0
astunparse=1.6.3=pypi_0
ca-certificates=2021.4.13=h06a4308_1
cachetools=4.2.2=pypi_0
certifi=2020.12.5=py38h06a4308_0
chardet=4.0.0=pypi_0
django=3.2.3=pypi_0
djangorestframework=3.12.4=pypi_0
gast=0.3.3=pypi_0
google-auth=1.30.1=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.38.0=pypi_0
h5py=2.10.0=pypi_0
idna=2.10=pypi_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.3.4=pypi_0
ncurses=6.2=he6710b0_1
numpy=1.20.3=pypi_0
oauthlib=3.1.0=pypi_0
opencv-python=4.5.2.52=pypi_0
openssl=1.1.1k=h27cfd23_0
opt-einsum=3.3.0=pypi_0
pandas=1.2.4=pypi_0
pillow=8.2.0=pypi_0
pip=21.1.1=py38h06a4308_0
protobuf=3.17.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
python=3.8.10=hdb3f193_7
python-dateutil=2.8.1=pypi_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scipy=1.4.1=pypi_0
setuptools=52.0.0=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.35.4=hdfb4753_0
sqlparse=0.4.1=pypi_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.2.0=pypi_0
tensorflow-estimator=2.2.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.10=hbc83047_0
urllib3=1.26.5=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3